feat: editando cabines

+++ src/services/apiCabins.js
@@ -13,34 +13,48 @@ export async function getCabins() {
     return data;
 }
 
-export async function createCabin(newCabin) {
-    const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "")
+export async function createAndEditCabin(newCabin, id) {
 
-    const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
+    const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
+    const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+        "/",
+        ""
+    );
+    const imagePath = hasImagePath
+        ? newCabin.image
+        : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
+    // 1. Create/edit cabin
+    let query = supabase.from("cabins");
 
-    // 1 - Create a new cabin
-    const { data, error } = await supabase
-        .from('cabins')
-        .insert([{ ...newCabin, image: imagePath }])
-        .select()
+    // A) CREATE
+    if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+    // B) EDIT
+    if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+    const { data, error } = await query.select().single();
 
     if (error) {
-        console.log(error)
-        throw new Error("Não foi possível criar a cabine");
+        console.error(error);
+        throw new Error("A cabine não pôde ser criada");
     }
 
-    // 2 - Upload the image
-    const { error: storageError } = await supabase.storage
-        .from('cabin-images')
-        .upload(imageName, newCabin.image)
+    // 2. Upload image
+    if (hasImagePath) return data;
 
-    // 3 - Delete the cabin IF there was an error uploading the image
+    const { error: storageError } = await supabase.storage
+        .from("cabin-images")
+        .upload(imageName, newCabin.image);
+
+    // 3. Delete the cabin IF there was an error uplaoding image
     if (storageError) {
-        await supabase.from('cabins').delete().eq('id', data.id)
-        console.error(storageError)
-        throw new Error("Não foi possivel fazer upload da imagem e a cabine não foi criada");
+        await supabase.from("cabins").delete().eq("id", data.id);
+        console.error(storageError);
+        throw new Error(
+            "Não foi possível fazer upload da imagem e a cabine não foi criada"
+        );
     }
 
     return data;
